requirements.txt
streamlit
faiss-cpu
pandas
numpy
sentence-transformers
transformers
scikit-learn
evaluate
torch
